pyproject.toml
[project]
name = "api-sk"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi==0.115.6",
    "passlib[bcrypt]==1.7.4",
    "pyjwt==2.10.1",
    "python-multipart==0.0.20",
    "uvicorn==0.34.0",
]


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
packages = ["api_sk"]
package-dir = {"" = "src"}

[project.scripts]
apisk = "api_sk.__init__:main"
